requirements.txt
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
Flask-Migrate==4.0.7
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
mysqlclient==2.2.4
playwright==1.47.0
prompt_toolkit==3.0.48
pyee==12.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
redis==5.1.0
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
